requirements.txt
Django==4.2.14
django-filter==23.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
pillow==10.4.0
pip==24.1.2
psycopg2==2.9.9
PyJWT==2.8.0
pyshorteners==1.0.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
